Combined coordinate, portal coordinate, and room coordinate into one coordinate file, and renamed files to follow effective dart styling guidelines
